move match functions to another file and add tests
