+++ frontend/src/app/utils/url.ts
@@ -28,3 +28,19 @@ export function getUrl(relative = false): string {
   }
   return "http://localhost:3000/";
 }
+
+export function stripIdTokenFromOktaRedirectUri(uri: string) {
+  const regexJWTAsQueryParam = /(?<=#id_token=).*?(?=&token_type=)/;
+  return stripIdTokenFromString(regexJWTAsQueryParam, uri);
+}
+
+export function stripIdTokenFromOperationName(operationName: string) {
+  const regexOperationName = /(?<=#id_token=).*/;
+  return stripIdTokenFromString(regexOperationName, operationName);
+}
+
+function stripIdTokenFromString(regex: RegExp, string: string) {
+  const idTokenFound = string.match(regex);
+  if (idTokenFound === null) return string;
+  return string.replace(idTokenFound[0], "{ID-TOKEN-OBSCURED}");
+}
